Finally input force field

+++ download_pdbs.py
@@ -5,12 +5,23 @@ import urllib
 import re
 from Bio.PDB import PDBParser, PDBIO, Select
 import shutil
+from torchmd_cg.utils.psfwriter import pdb2psf_CA
+import argparse
+
+def get_args(arguments=None):
+    parser = argparse.ArgumentParser(description='TorchMD-AD-download', prefix_chars='--')
+    parser.add_argument('--pdbs', default=None, help='Download PDBs')
+    parser.add_argument('--psfs', default=None, help='PDBs to PSFs')
+    
+    args = parser.parse_args(args=arguments)
+
+    return args
 
 dms_dir = os.path.dirname(os.path.realpath(__file__))
 dataset_dir = os.path.join(dms_dir, "datasets")
 train_val_dir = os.path.join(dms_dir, "protein_data", "train_val")
 pdbs_dir = "/workspace7/torchmd-AD/train_val_pdb/train"
-
+psfs_dir = "/workspace7/torchmd-AD/train_val_psf/train"
 
 # Extract the pdb and chain names of a file 
 def extract_pdb_code(name_w_chain):
@@ -85,7 +96,9 @@ def extract_chains(pdb_chains, input_path, output_path):
                         io.set_structure(chain2)
                         io.save(output_path + protein + '_' + chain2.get_id() + ".pdb", NonHetSelect())
 
-    
+
+                   
+                          
 class NonHetSelect(Select):
     """ Avoid selecting heteroatoms from a structure"""
     
@@ -95,20 +108,30 @@ class NonHetSelect(Select):
 
 
 if __name__ == "__main__":
+    args = get_args()
     
-    train_proteins = [l.rstrip() for l in open(os.path.join(dataset_dir, "train.txt"))]
+    if args.pdbs:
+        train_proteins = [l.rstrip() for l in open(os.path.join(dataset_dir, "train.txt"))]
     
-    # Download all the files in pdb format
-    tmp_dir = pdbs_dir + 'downloads/'
-    os.mkdir(tmp_dir) # tmp directory to save full pdbs
+        # Download all the files in pdb format
+        tmp_dir = pdbs_dir + 'downloads/'
+        os.mkdir(tmp_dir) # tmp directory to save full pdbs
     
-    for file in train_proteins:
-        download_pdb(file, pdbs_dir + 'downloads/')
+        for file in train_proteins:
+            download_pdb(file, pdbs_dir + 'downloads/')
     
-    # Dictionary with proteins and their chains
-    pdb_chains = pdb_chain_to_dict(train_proteins)    
+        # Dictionary with proteins and their chains
+        pdb_chains = pdb_chain_to_dict(train_proteins)    
     
-    # Extract desired chains from downloaded proteins
-    extract_chains(pdb_chains, pdbs_dir + 'downloads/', pdbs_dir)
+        # Extract desired chains from downloaded proteins
+        extract_chains(pdb_chains, pdbs_dir + 'downloads/', pdbs_dir)
     
-    shutil.rmtree(tmp_dir) # rm tmp directory
+        shutil.rmtree(tmp_dir) # rm tmp directory
+    
+    elif args.psfs:
+        # Transform the pdbs to psfs
+        for file in os.listdir(pdbs_dir):
+            PDB_file = pdbs_dir + '/' + file
+            PSF_file = psfs_dir + '/' + file[:-4] + '.psf'
+            pdb2psf_CA(PDB_file, PSF_file)
+            # TODO: SOLVE RUNTIME ERROR
